+++ display.h
@@ -4,7 +4,7 @@
 #include "matrix.h"
 
 void init_z_buffer( struct matrix* zb );
-void plot( screen s, color c, int x, int y, int z, struct matrix *zb );
+void plot( screen s, color c, int x, int y, double z, struct matrix *zb );
 void clear_screen( screen s);
 void save_ppm( screen s, char *file);
 void save_extension( screen s, char *file);
